Fix state of charge calculation
The CalcSOC() function was not being called in the
Ms100Task though it looks like it is supposed to.

All of the functions are now included automatically in the
"utils" namespace in a more consistent manner to ensure
that everthing ends up in the correct namespace and not
accidentally in the global namespace.

Tests:
 - Build and vierfy that the warning about the use of
   utils::CalcSOC has gone.