Change i3LIM Chg_Phase from raw values to an enum
The charge phase has a lot of different values. Make things
clearer by changing to an enum with those values and
remove the now duplicate comments. Use the term
"Supoena" rather than "PreCharge" as this seems to be
more common in the existing code. Use "Shutdown" rather
than "Turn off" as this seems to match the comments better.