i3LIM: Give contactors time to close before starting energy transfer
During the ChargePhase::Subpoena (pre-charge) we wait
for the contactor voltage to get within an acceptable range.
Once this happens the contactors are requested to close.
We should give time for the contactors to close before
moving to the ChargePhase::EnergyTransfer.

There is a possibility during lim_state 4 that the contactor
voltage drifts outside the acceptable range before we close
the contactors. If this happens reset the timer until it comes
back inside the range.