i3LIM:Proceed to Energy Transfer when contactors report closed
Report the CCS contactor state as a read-only parameter.
Use the change in CCS contactor state to advance out of
Subpoena/pre-charge and into the EnergyTransfer phase
rather than wait for a fixed period of time.

This fixes a bug where the code would get stuck in lim_state
== 5 because lim_stateCnt was not being incremented so
the timer waiting for contactors to close never completed.

Tests:
 - Compile tested only