Fix punctuation of enum parameters
There are punctuation errors in the units for the
"Inverter", "chargemodes", "CAN3Speed" and
"OutXFunc". The first 3 cause parse errors for the
openinverter can tool when trying to read the
parameter database from the device as covered in
https://github.com/davefiddes/openinverter-can-tool/issues/4.
Whilst this can be worked around client side it is
better fixed in the Zombieverter.

The issue likely causes problems with the web interface missing one or more items from the drop down.

Tests:
 - Compile tested only
 - Verify a fixed database is parsed correctly by OIC