Fix linker warning from newer toolchains
Import commit 53afdb04 from stm32-sine which
suppresses a linker warning:

/usr/lib/gcc/arm-none-eabi/14.1.0/../../../../arm-none-eabi/bin/ld: warning: stm32_vcu has a LOAD segment with RWX permissions

From the original commit:
Due to the way that gcc generates .init_array sections
containing C++ global object constructors it can result
in ELF object file sections that are marked Read Write
and eXecute. With binutils 2.39 or newer this will
result in a warning. For an embedded use case like
this inverter it can safely be suppressed.

Suppressing the warning is not possible on earlier
versions of binutils (GNU ld). The makefile now tests
whether the link option will succeed before trying to
use it.

Tests:
 - Build test with gcc 14.1.0 and binutils 2.43