Add Continuous Integration build using GitHub Actions
There has been a lot of accidental build breakage with
recent commits. To reduce the likelihood of this in
future ensure that pushes and PRs are tested by
running a clean build in a known clean environment
and report any errors.

This uses the same approach as the stm32-sine and
ccs32clara projects for continuous integration.

The readme is updated to display the current CI build
state. It has also been update to reflect the changed
build process for unit tests.