Make GS450H data tables constant to save RAM
The initial set up data for GS450H, GS300 and Prius
are never modified and only used to send data to the
inverter. They can be defined as const which saves
RAM and time to copy them from Flash at startup.

The dma_write function is updated to take a const
pointer. This is safe as it is being transmitted via DMA
and not modified as it would be for receive.

Tests:
 - Compile test only