Fix KangoosBMS interface
Update the includes in kangoobms.h so that they are
sufficient to define the KangooBMS class. Also
update kangoobms.cpp to minimise the number of
includes used.

Mark overridden virtual methods with "override" to
reduce likelihood of future breakage should the
parent method signature change.

Remove the GetCurrent() method from the interface
as there is no straightforward way for the rest of the
project to call this and it is otherwise unused.

Tests:
 - Compile test only