Simplify i3LIMClass interface
Remove the static designator from private methods in
i3LIMClass. This gives the optimizer more
opportunities to inline methods. The data should
probably be moved to be class members but that is
too intrusive at this time.

Mark overridden virtual methods with "override" to
reduce likelihood of future breakage should the
parent method signature change.

Tidy the includes.