Simplify CAN callbacks to use bytes rather than 32-bit words
Virtually all CAN callbacks cast the two 32-bit words
the libopeninv library presents an incoming CAN
frame in to bytes. Rather than repeat this cast all over
the place replace with one central cast and change
the interface of the DecodeCAN callback methods.

The exceptions which require reverse casts of byte-
arrays back to 32-bit words are:
 - FCChademo
 - E65_Lever
 - OutlanderInverter

These could be changed to access the bytes directly but
this would require greater understanding and testing of
the code.

The changes are mechanical in nature and should not
affect the operation of the code.

Where applicable "override" is used for virtual methods
to make it clear to the compiler that virtual functions
are being overridden.

Tests:
 - Build for STM32 and Linux