Fix TorqDerate and CAN3Speed enum definitions
The TorqDerate spot value does not display properly
in the esp8266-web-interface. With a TorqDerate of
13 it displays UDClimHigh. The problem is due to "+"
being an invalid character for enum value names. As
the field is used as a bitfield change the definition to a
bitfield enum which the web interface can display.
Verify behaviour with power-of-two values and mix of
values between 0 and 31.

Refix punctuation of CAN3Speed enum to allow it to
be parsed by openinverter-can-tool