HVIL
Added a HV interlock input, grounded at one end and into a zombie PB input on the other. Or can be 12v into a 12v input I guess.
In the event of the interlock being broken, HVIL will throw an error but not  affect anything during run or charge mode. But it will prevent precharge from happening.